requirements.txt
asgiref==3.5.2
Django==4.0.6
django-environ==0.9.0
django-mysql==4.7.0
djangorestframework==3.13.1
mysqlclient==2.1.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
